Improve the size computation for txb buffers
The txb buffers are allocated in av1_alloc_txb_buf() based on
the number of superblocks present in a frame. This was computed by
dividing mi_rows and mi_cols with the superblock size in MI units
and then adding a factor of 1 to both in order to account for
rounding errors. This patch improves this allocation by using a
logic similar to ceil() so that no additional memory is allocated
when either of the two dimensions is a perfect multiple of the
superblock size.

For AVIF image encode with speed = 9,

             HEAP Memory reduction(%)
Resolution   threads=1    threads=4
640x360        1.77         1.47
768x512        7.83         6.37
832x480        1.54         1.40
1280x720       2.32         2.04

HEAP memory reduction was measured using the following command.
$valgrind --tool=massif ./avifenc ...

Change-Id: Ic91bf838498fa76f24fe32f6d6d23c7654d4a0cf