AV1 RT: Implement palette mode for nonRD path
Initial implementation only checks luma palette. It kicks in only if
tune-content is set to screen and only if intra mode search did not lead
to skip transform and only on scene change.

Results on screen content with speed 8 and on 500k
3-10% slower and 0.6-1.6 dB better. On scene change the perf impact
is ~2x slower. Visually the difference is very noticeable on some clips.

Change-Id: I206d21aad4ee545d5defd2cfa45b35f37ac60a08