Reduce arnr-strength to 1 during keyframe filtering mode 1
This fixes https://bugs.chromium.org/p/aomedia/issues/detail?id=3211
aka the blocking issue that made `--enable-keyframe-filtering=1`
unusable.

The compression improvement is not as large as with the full ARNR value,
but it is better than disabling keyframe filtering. Ideally, I think the
proper fix is to change to a different filtering method that doesn't
temporally destroy frames. But that is a significant amount of work.