Avoid abs_sum_level computation conditionally
abs_sum_level is needed to decide the job scheduling order of
pack bitstream multi-threading. The computation can be avoided
when MT is disabled or multiple tiles are not used. This change
could avoid additional overhead due to computation of
abs_sum_level.

Speed data for RT preset.
     Instruction Count
cpu    Reduction(%)
 7       0.189
 8       0.236
 9       0.434
10       0.474

Change-Id: Id83f97b023f6e5275d1bfd59b889bcfb99dee067