Modify and simplify propagation mechanism
The propagation mechanism in TplFrameDepStatsPropagate()
is modified based on the paper
"A Temporal Dependency Model for Rate-Distortion Optimization in
Video Coding".

We also simplify TplFrameDepStatsPropagate() by converting
TplBlockStats to TplUnitDepStats before hand so that TplFrameStats
is not needed in TplFrameDepStatsPropagate().

We also make sure inter_cost <= intra_cost when converting
TplBlockStats to TplUnitDepStatsin by TplBlockStatsToDepStats()

Bug: b/221916304
Change-Id: Iafc2af5557fa99977ba82e1cf8b45b5863c78b6f