Use dedicated Arm hardware instructions to compute CRC32C hash
Add an implementation of the CRC32C hash function that uses the
dedicated CRC32C instructions in the Armv8-A architecture.

The Arm CRC32C instructions are optional in Armv8.0-A but mandatory
in all architecture revisions starting from Armv8.1-A. For the time
being, we detect the presence of the CRC32C instructions at compile
time but run-time feature detection ought to be implemented for
platforms like Android where an architecture baseline of Armv8.1-A
cannot be assumed.

Change-Id: I30894ec77dd79c87eeafcdefc9e4b061cd5106d6