rtc-screen: Allow DC intra for large blocks
Current max block size for Intra check was 32x32,
but increase it max(128x128) for screen, but only
for DC mode and only when spatial variance is 0 and
only on slide changes.

Removes bad artifact on slide change when flat area
changes only by the dc shift.

Negiligible change in bdrate and speed, as the condition
does not trigger often.

Change-Id: I378d5119c5e42e4ef718cc1206c6eb2e7a849abd