rtc-screen: Fix to skipping inter-mode test in nonrd
For spatially flat blocks with non-zero motion
inter-modes with zero motion were skipped. But
this logic is done before newmv search, and so
if the newmv motion vector is (0, 0), then an
assert will be triggered in encode_mv if nearestmv
or nearmv are also (0, 0).

Removing that skip condition for zero motion fixes
the issue; also the condtion is basing the non-zero
motion on superblock sad, instead of coding block,
so not accurate/reliable and so better to remove it
for now.

This only affects screen-content mode.
Overall neutral on bdrate and speed for speed 10.

Add test with many scene cuts and high motion,
this triggers the assert in the issue without this fix.

Bug: aomedia:3289
Change-Id: Ib52c67d22ccf66b39a43c449b8edc91204917017