Make number_spatial_layers_ a local var in RunLoop
Convert EncoderTest::number_spatial_layers_ to a local variable in
EncoderTest::RunLoop().

The number_spatial_layers_ data member is assigned a value in
EncoderTest::RunLoop() and then only used in that method, so it is
effectively a local variable in that method.

NOTE: This also fixes the problem that both the base class EncoderTest
and the derived class DatarateTestSVC have an
"int number_spatial_layers_" member. Although C++ allows this, this is
very confusing.

Also fix an off-by-one error in a comparison in the for loop. The loop
index variable `sl` is never equal to number_spatial_layers_.

Ref: https://aomedia-review.googlesource.com/c/aom/+/93902

Change-Id: Ie3b325a64c424561f1df441077279ed9bfdc805b