rtc-screen: Avoid source_variance computation
For screen mode force_zeromv_skip may be set,
which means the pickmode is effectively skipped
(zeromv-last is forced), so no need to compute
the source_variance when this is set. This saves
a variance computation for the coding block.

Bitexact change: ~1 IC (instruction count)
speedup on speed 10 rtc_screen.

Change-Id: I2dd25d6f81bd31348aa11cd1524c73f39fee5781