rtc: Fix to disable segm for aq-mode=3
Disable the segmentation if the target_num_seg_blocks
is 0, this can happen for content with high motion
(where cr->map may be 1 for all blocks on some frames).
Disabling the segmentation (seg_map is set to 0 at start
of the function, so no need to reset) reduces overhead and
gives small bdrate gain on a few clips (others have no change).

For speed 10 rtc_derf, the only clips with change:
apprtc, still_bright, street, and room_scene_cut, which get:
0.32, 0.452 0.15, 0.57 % bdrate gain.

Change-Id: I5ddcf9654ed884d8d9682998424bf94c4061fbf5