Do not skip updating active map when it is disabled
Currently, in av1_apply_active_map function, active_map is not updated
when it is disabled. But this causes the update flag to be kept as 1 in
active_map. Consequently, when suppress_active_map is called, all
segments are marked as active map. For realtime mode, this leads to over
0.5% of encoding time increase.

This commit removes the code that skips the active map update to resolve
this issue.

Change-Id: Ib2b17f45f1d76889c7096159ed34615a007170af