Simplify partition_search_breakout_dist_thr related code
partition_search_breakout_dist_thr threshold/speed feature
was corresponding to a block of size 64x64. Hence this
threshold was multiplied by 4 in
av1_set_speed_features_framesize_independent() to map to
maximum block size of 128x128. However, similar scaling
was not present in
av1_set_speed_features_framesize_dependent(). In this CL,
the scaling code is removed and the thresholds are modified
based on the resolution for speed=5 and 6 presets to
account for multiplication by 4.

STATS_CHANGED for --good usage

Change-Id: I02110c43e7f0908c7312e24953852a036b8988e3