rtc: Allow for setting 1 layer reference struct externally
The control SET_SVC_REF_FRAME_CONFIG allows for the application
to set the full reference structure. But for single layer
encoding we would still default to the internal reference
structure define in av1_set_reference_structure_one_pass_rt().

This change allows the user/application to also define the
reference structure for single layer (non-svc) encoding via
the same control. If the user does not use the control then
the default internal structure will be used. This is useful
to give more control to the app and also for the usage of
RPS (reference picture selection) for 1 layer.

The relevant functions are also renamed to make this clear.

Also renaming the control SET_SVC_REF_FRAME_CONFIG to
SET_RTC_REF_FRAME_CONFIG and related refactoring is considered in:
https://aomedia-review.googlesource.com/c/aom/+/161827

No change in stats.
Change-Id: Iead7c028d14b87a17588efb6c3447c28e4eba180