Add mixed-sign dot-product 12-tap convolution optimizations
Add an additional Neon path for av1_convolve_2d_sr_horiz_12tap_neon
that makes use of the Armv8.6-A USDOT instruction.

USDOT takes an 8-bit unsigned operand vector and a signed 8-bit
operand vector to produce a signed 32-bit result. This is helpful
because convolution filters often have both positive and negative
values, while the 8-bit pixel channel data being filtered is all
unsigned. As a result, the USDOT convolution paths added here do not
have to do the "transform the pixel channel data to [-128, 128) and
correct for it later" dance that we have to do with the SDOT paths.

The USDOT instruction is optional from Armv8.2 to Armv8.5 but
mandatory from Armv8.6 onwards. The availability of the USDOT
instruction is indicated by the feature macro
__ARM_FEATURE_MATMUL_INT8. The SDOT paths are retained for use on
target CPUs that do not implement the USDOT instructions.

Change-Id: I3bef9738e382baaf1c956e3a5682195c435e36e8