Add mixed-sign dot-product 8-tap convolution optimizations
Add additional Neon paths for the following convolution functions:
 * av1_convolve_2d_sr_horiz_neon
 * dist_wtd_convolve_2d_horiz_neon
 * av1_convolve_x_sr_neon
 * av1_dist_wtd_convolve_x_neon

The new paths make use of the Armv8.6-A USDOT instruction which can
take an 8-bit unsigned operand vector and a signed 8-bit operand
vector to produce a signed 32-bit result. This is helpful because
convolution filters often have both positive and negative values,
while the 8-bit pixel channel data being filtered is all unsigned. As
a result, the USDOT convolution paths added here do not have to do
the "transform the pixel channel data to [-128, 128) and correct for
it later" dance that we have to do with the SDOT paths.

The USDOT instruction is optional from Armv8.2 to Armv8.5 but
mandatory from Armv8.6 onwards. The availability of the USDOT
instruction is indicated by the feature macro
__ARM_FEATURE_MATMUL_INT8. The SDOT paths are retained for use on
target CPUs that do not implement the USDOT instructions.

Change-Id: I66a2dda6026ccd7efdcd323e739fe9751eb9973d