Refactor loopfilter_frame()
In this CL, the default initialization of the in-loop
post-processing filter parameters is abstracted to a function
set_inloop_postproc_filter_default_params(). This CL also
abstracts the code to extend the frame borders into a function
extend_frame_borders() and its call is moved to facilitate
the subsequent CL which conditionally skips applying in-loop
post-processing filters on a reconstructed frame. This is a
bit-exact change.

Change-Id: Ifcde3faf9a3c4f5e025ef417685e0d20ae8d5db1