rtc: Fix mv signaling when mv is same as refmv
If NEWMV motion vector is equal to that of NEARESTMV,
it is intended that we signal NEARESTMV. This is
enforced by an assert in av1_encode_mv(). In the
parent version, assert is hit for a testcase, in
which all of the below are true for an inter block:
1. {ALTREF, NEWMV} is chosen as the final pred mode
2. {ALTREF, NEWMV} is same as {ALTREF, NEARESTMV}
3. {ALTREF, NEARESTMV} is not evaluated due to an
early term condition

In this CL, we fix this issue by adding a condition
in search_new_mv(). For NEARMV, this is not needed
since early-term of NEARMV would mean early-term of
NEWMV as well.

The Borg test results are:
                   Instr Cnt        BD-Rate Loss(%)
cpu      Testset    Red (%)  avg.psnr  ovr.psnr    ssim
 7           rtc    0.042     0.0012    0.0007    0.0013
 7     rtc_1080P    0.029     0.0000    0.0000    0.0000
 8           rtc    0.050     0.0096    0.0093    0.0159
 8     rtc_1080P    0.036     0.0000    0.0000    0.0000
 9           rtc    0.026     0.0000    0.0000    0.0000
 9     rtc_1080P    0.017     0.0018    0.0018    0.0052

No changes to other cpu and rtc_derf/rtc-screen testsets.

STATS_CHANGED for rt

Change-Id: I87548f92d6f4ff1e9c085571067e89d17d5413dc