rtc-screen: Fix Q setting for forced key frames
If the user in RTC sets/forces a key frame in the
stream (this_key_frame_forced = 1) the q for that
frame will be the last_boosted_qindex which can be
very low, especially for screen, and cause very
large encoded key frame size.

This logic shouldn't be there for CBR, but for now
disable it for screen only, where excessive frames size
can occur due to this.

This reduces the key frame size by ~3-20x (on slidechanges clip,
depending on the scene change); quality is reduced, but gets
cleaned up quickly with aq-mode=3.

No change in stats for rtc sets (as no forced key frames
are set by default).

Change-Id: I606cfe617698e09859231b0bb80b5b7ae212737e